centralize report code
